requirements.txt
dill==0.2.6
flock==0.1
keras==2.0.3
matplotlib==2.0.2
numpy==1.13.0
opencv_python==3.2.0.7
pandas==0.20.2
pathos==0.2.0
Pillow==4.1.1
scikit_image==0.13.0
scipy==0.19.0
Shapely==1.5.17.post1
tensorflow_gpu==1.2.0
